feat: rewrite types
