+++ loxrs_entities/Cargo.toml
@@ -6,3 +6,5 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+loxrs_types = {path = "../loxrs_types"}
+loxrs_env = {path = "../loxrs_env"}
