+++ loxrs_entities/src/lib.rs
@@ -1,3 +1,12 @@
-pub mod error;
+pub mod expr;
+pub mod stmt;
+pub mod token;
+pub mod token_type;
+pub mod val;
 
-pub use error::{LoxErr, Result};
+pub use expr::Expr;
+pub use stmt::Stmt;
+pub use token::Token;
+pub use token_type::TokenType;
+pub use val::Literal;
+pub use val::Value;
