+++ loxrs_interpreter/Cargo.toml
@@ -10,3 +10,4 @@ env_logger = "0.11.3"
 log = "0.4.21"
 loxrs_entities = {path = "../loxrs_entities"}
 loxrs_env = {path = "../loxrs_env"}
+loxrs_types = {path = "../loxrs_types"}
