--- loxrs_interpreter/src/lox/interpreter/entities.rs
@@ -1,13 +1,0 @@
-pub mod expr;
-pub mod fns;
-pub mod stmt;
-mod token;
-mod token_type;
-mod val;
-
-pub(super) use expr::Expr;
-pub(super) use stmt::Stmt;
-pub(super) use token::Token;
-pub(super) use token_type::TokenType;
-pub(super) use val::Literal;
-pub(super) use val::Value;
