+++ loxrs_types/Cargo.toml
@@ -1,11 +1,8 @@
 [package]
-name = "loxrs_env"
+name = "loxrs_types"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-loxrs_types = {path = "../loxrs_types"}
-env_logger = "0.11.3"
-log = "0.4.21"
