+++ loxrs_types/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod error;
+pub use error::{LoxErr, Result};
